requirements.txt
certifi==2019.9.11
chardet==3.0.4
DBUtils==1.3
idna==2.8
mysql-connector-python==8.0.17
protobuf==3.9.1
psycopg2-binary==2.8.3
pymongo==3.9.0
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
urllib3==1.25.3
